Move inherits function to schemas
